fix(RSS-ECOMM-3_08): build fixes
